refactor(modularization): Phase 4 Step 4-S3c — marketplace schemas extract (SharedCollection* → aperag/domains/marketplace/schemas.py + dual-hook) (#25)
Third sub-step of 4-S3. 3 schemas carved to marketplace domain:
SharedCollectionConfig / SharedCollection / SharedCollectionList.

Boundary tests 15/15 green. Dual-hook re-import verified via
view_models try block.
